+++ UWRandomizerEditor/UWRandomizerEditor.csproj
@@ -15,10 +15,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Settings.cs"/>
-  </ItemGroup>
-
-  <ItemGroup>
     <None Update="UWRandomizerEditor.dll.config">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
